+++ CHANGELOG.md
@@ -19,8 +19,8 @@ Possible log types:
 - [added] `heap` option for `heap!` macro has been renamed to `metadata`
 - [added] Added `global` option to `heap!` macro
 - [changed] Changed function signature for outer thread handlers
-- [changed] `fib::Chain::drain` method now returns a `bool`, which indicates
-  whether any fiber was executed.
+- [changed] `fib::Chain::drain` method now returns an iterator, which can be
+  queried whether any fiber was executed.
 - [added] Added a new synchronization primitive: lock-free singly-linked list
   `sync::LinkedList`
 - [changed] `sync::Mutex` was rewritten to be futures-aware